extract state function sFnAddFinalizer from sFnInitialize
